Move api-client.tsx to utils
