+++ frontend-streamhub-main/src/utils/api-client.tsx
@@ -1,9 +1,9 @@
-import { LoginFormData } from "./pages/LoginPage";
-import { RegisterFormData } from "./pages/RegisterPage";
+import { LoginFormData } from "../pages/LoginPage";
+import { RegisterFormData } from "../pages/RegisterPage";
 import axios from "axios";
 import { User } from "@/utils/types";
-import { UpdateFormData } from "./pages/UpdateProfilePage";
-import { useAppContext } from "./contexts/AppContext";
+import { UpdateFormData } from "../pages/UpdateProfilePage";
+import { useAppContext } from "../contexts/AppContext";
 
 export const register = async (formData: RegisterFormData) => {
   const response = await axios
